==> 08_stat_models.rmd <==
---
output: html_document
editor_options: 
  chunk_output_type: console
---

# Statistical modelling

```{r}
# to handle data
library(readr)
library(scales)
library(tidyr); library(tibble)
library(magrittr)
library(dplyr)
library(purrr)
library(stringr)
library(forcats)

# to plot
library(ggplot2)
library(scico)

# functions
ci <- function(x){qnorm(0.975)*sd(x, na.rm = T)/sqrt((length(x)))}
```

## Read data

```{r}
read_csv(data_summary, path = "data/data_pairwise_comparisons.csv")
```

### Subsample for even coverage

Small gizzard mass differences are over-represented as expected from a population with normally distributed traits. This requires subsampling based on difference bins.

```{r}
# filter for differences above 5 and subsample as many
# as supported
data_subsample <- data_summary %>% 
  filter(diff_gizzard_round <= 3) %>% 
  group_by(tide_stage, diff_gizzard_round) %>% 
  nest() %>%
  mutate(data = map(data, function(x) x[1:1480,])) # this is hardcoded

# unnest
data_subsample <- unnest(data_subsample, cols = "data")
```

## Statistical modelling of number of associations

### Fit a GLMM for number of associations

Difference in gizzard mass and behaviour score are predictors.

```{r}
# run a single canonical model
library(lmerTest)
model_n_assoc_full <- glmer(n_associations ~ diff_gizzard * tide_stage +
                         diff_behav * tide_stage +
                         (1|tide_number) +
                         (1|uid_pair),
                       data = data_subsample,
                       family = "poisson")

# run an alternative model with only behaviour
model_n_assoc_behav_only <- glmer(n_associations ~ 
                                 diff_behav * tide_stage +
                                 (1|tide_number) +
                                 (1|uid_pair),
                               data = data_subsample,
                               family = "poisson")

# an alternative model with only gizzard mass
model_n_assoc_gizzard_only <- glmer(n_associations ~ 
                                 diff_gizzard * tide_stage +
                                 (1|tide_number) +
                                 (1|uid_pair),
                               data = data_subsample,
                               family = "poisson")
```

### AIC comparison

Compare the AIC weights after looking at the model summaries.

```{r}
summary(model_n_assoc_full)

summary(model_n_assoc_behav_only)

summary(model_n_assoc_gizzard_only)
```

It seems that differences in behaviour as well as gizzard mass are each sufficient sufficient to explain the observed number of overlaps. Examine the AIC scores.

```{r}
# get aic difference
map_dbl(list(behaviour_only = model_n_assoc_behav_only, 
         gizzard_only = model_n_assoc_gizzard_only, 
         gizzard_and_behav = model_n_assoc_full),
    AIC)
```

With similar AIC scores and R<sup>2</sup> values, we select the reduced model by the parsimony principle, and write the model output to file.

```{r}
model_output = capture.output(map(list(behaviour_only = model_n_assoc_behav_only, 
                                       gizzard_only = model_n_assoc_gizzard_only, 
                                       gizzard_and_behav = model_n_assoc_full),
                                  summary))

# save it to file
write_lines(x = model_output, path = "results/model_n_assoc.txt")
```

Read in the model and print summary.

```{r}
cat(read_lines("results/model_n_assoc.txt"), sep = "\n")
```


## Statistical modelling of spatial overlap

### GLMM for spatial overlap

```{r}
# run a model with overlap as the response
library(lmerTest)
model_sp_overlap_full <- lmer(tot_overlap_space ~ diff_gizzard * tide_stage +
                         diff_behav * tide_stage +
                         (1|tide_number) +
                         (1|uid_pair),
                       data = data_subsample)

# run an alternative model with only behaviour
model_sp_overlap_behav_only <- lmer(tot_overlap_space ~ 
                                 diff_behav * tide_stage +
                                 (1|tide_number) +
                                 (1|uid_pair),
                               data = data_subsample)

# an alternative model with only gizzard mass
model_sp_overlap_gizzard_only <- lmer(tot_overlap_space ~ 
                                 diff_gizzard * tide_stage +
                                 (1|tide_number) +
                                 (1|uid_pair),
                               data = data_subsample)
```

### AIC comparison

Compare the AIC weights after looking at the model summaries.

```{r}
summary(model_sp_overlap_full)

summary(model_sp_overlap_behav_only)

summary(model_sp_overlap_gizzard_only)
```

It seems that differences in behaviour are sufficient in this case to explain the observed extent of overlaps. Examine the AIC scores.

```{r}
# get aic difference
map_dbl(list(behaviour_only = model_sp_overlap_behav_only, 
         gizzard_only = model_sp_overlap_gizzard_only, 
         gizzard_and_behav = model_sp_overlap_full),
    AIC)
```

```{r}
model_output = capture.output(map(list(behaviour_only = model_sp_overlap_behav_only, 
                                       gizzard_only = model_sp_overlap_gizzard_only, 
                                       gizzard_and_behav = model_sp_overlap_full),
                                  summary))

# save it to file
write_lines(x = model_output, path = "results/model_sp_overlap.txt")
```

Read in the model summary and write.

```{r}
cat(read_lines("results/model_sp_overlap.txt"), sep = "\n")
```
